requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
requests-mock>=1.11.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.8.0
